Add simulated movement
Now player positions are updated constantly, providing smooth movement. We are
now sending the whole game state for simplicity. In the future we can optimise
this. I was thinking of maybe using an entity system where everything in the
game state has an ID and the server only sends the states of entities that have
changed (and sending nothing if nothing has changed).